backend/requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
python-multipart
python-jose[cryptography]
passlib[bcrypt]